finsentinel-backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo[srv]
certifi
python-dotenv
pydantic
scikit-learn
joblib
pandas
numpy